Move add-extension button to gui.jsx
